+++ client/playlist/playlist.js
@@ -5,7 +5,17 @@ Template.playlist.helpers({
     },
     durationDisplay: function() {
     	return timeFormat(this.duration, this.duration);
-   	}
+   	},
+    playlistDuration: function() {
+      var Songs = Playlist.find({'position': {$gte: 1}});
+      // the total duration
+      var totalDuration = 0;
+      Songs.forEach(function (song) {
+        totalDuration += song.duration;
+      });
+      // return the formatted timestring
+      return timeFormat(totalDuration, totalDuration);
+    }
 });
 
 Template.playlist.events({